Commited some changes to the design of the website
After a little brainstorming session, i just decided to change the overall design of the website. I have uploaded a new color palet readme file to give more insight of the brand new colors that will be used. The colors are implemented to all the HTML pages of the website.