Commited some changes to the folder structure
Just changed the folder name from "Privé leven" to "over-mij". I also changed all the names on the website from "Privé leven" to "Over mij". This includes changes to the pathnames in the navigation bar and under every tiles that i created on the Index, Over mij en Traject school pages. This to create a better best practice for file and folder names structures.