Commit version (v0.50.0) for Project1
This commit applies the final touches across all pages. Version 0.50.0 marks the end of Project 1 and the beginning of Project 2.

Changes
Updated the images used throughout the site.
Changed the text color from black to white on the remaining pages and changed some headers to be more inline with the theme of the website.

Note
Due to time constraints, a few responsiveness fixes have not yet been applied.